Create Separator, CampusMenuInfoView files; Fix constraint conflict
